+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+## v0.4.7
+- Fixed exception when try to insert in horus_entities with a entity name exists.
+- Fixed issue in Entity get attributes polymorphic like getInt and getFloat.
+
 ## v0.4.6
 - Fixed issue when it tries to insert or update a entity with a null value.